Break pixel-sortying logic out into a utils folder to clean up CreatePage
