+++ server/data/7f3e6a81-129a-44c1-a0c6-2b1e67b8f747-settings.json
@@ -1,0 +1,6 @@
+{
+  "sortingThreshold": 120,
+  "colorChannel": "green",
+  "sortingDirection": "vertical",
+  "timestamp": "2024-02-04T18:04:01.848Z"
+}
